refactor: extract source tag to own component
